Estimation

Procedure

The goal of the estimation step is to sample from the posterior distribution of the model parameters. DSGE.jl uses a Metropolis-Hastings sampler to do this, which requires as a proposal covariance matrix the Hessian matrix corresponding to the posterior mode. The function estimate implements the entire procedure.

Main Steps:

Initialization: Read in and transform raw data from save/input_data/. See Input Data for more details.

Reoptimize parameter vector: The main program will call the csminwel optimization routine (located in csminwel.jl) to find modal parameter estimates.

Compute Hessian matrix: Computing the Hessian matrix to scale the proposal distribution in the Metropolis-Hastings algorithm.

Sample from Posterior: Posterior sampling is performed using the Metropolis-Hastings algorithm. A proposal distribution is constructed centered at the posterior mode and with proposal covariance scaled by the inverse of the Hessian matrix. Settings for the number of sampling blocks and the size of those blocks can be altered as described in Editing or Extending a Model.

Remark: In addition to saving each mh_thin-th draw of the parameter vector, the estimation program also saves the resulting posterior value and transition equation matrices implied by each draw of the parameter vector. This is to save time in the forecasting step since that code can avoid recomputing those matrices.

To run the entire procedure, the user simply calls the estimate routine:

DSGE.estimate — Function.

estimate(m, data; verbose=:low, proposal_covariance=Matrix())

Estimate the DSGE parameter posterior distribution.

Arguments:

m::AbstractModel: model object

Optional Arguments:

data: well-formed data as Matrix or DataFrame. If this is not provided, the load_data routine will be executed.

Keyword Arguments:

verbose::Symbol: The desired frequency of function progress messages printed to standard out.

:none: No status updates will be reported.

:low: Status updates will be provided in csminwel and at each block in Metropolis-Hastings.

:high: Status updates provided at each iteration in Metropolis-Hastings.

proposal_covariance::Matrix: Used to test the metropolis_hastings algorithm with a precomputed covariance matrix for the proposal distribution. When the Hessian is singular, eigenvectors corresponding to zero eigenvectors are not well defined, so eigenvalue decomposition can cause problems. Passing a precomputed matrix allows us to ensure that the rest of the routine has not broken.

Computing the Posterior

In DSGE.jl, the function posterior computes the value of the posterior distribution at a given parameter vector. It calls the likelihood function, which in turn calls the kalman_filter routine. See Estimation routines for more details on these functions.

We implement the Kalman Filter via the kalman_filter function to compute the log-likelihood, and add this to the log prior to obtain the log posterior. See State Space Routines for a model-independent implementations of the Kalman filter.

Optimizing or Reoptimizing

Generally, the user will want to reoptimize the parameter vector (and consequently, calculate the Hessian at this new mode) every time they conduct posterior sampling; that is, when:

the input data are updated with a new quarter of observations or revised

the model sub-specification is changed

the model is derived from an existing model with different equilibrium conditions or measurement equation.

This behavior can be controlled more finely.

Reoptimize from a Specified Starting Vector

Reoptimize the model starting from the parameter values supplied in a specified file. Ensure that you supply an HDF5 file with a variable named params that is the correct dimension and data type.

m = Model990()
params = load_parameters_from_file(m, "path/to/parameter/file.h5")
update!(m, params)
estimate(m)

Skip Reoptimization Entirely

You can provide a modal parameter vector and optionally a Hessian matrix calculated at that mode to skip the reoptimization entirely. These values are usually computed by the user previously.

You can skip reoptimization of the parameter vector entirely.

m = Model990()
specify_mode!(m, "path/to/parameter/mode/file.h5")
estimate(m)

The specify_mode! function will update the parameter vector to the mode and skip reoptimization by setting the reoptimize model setting. Ensure that you supply an HDF5 file with a variable named params that is the correct dimension and data type. (See also the utility function load_parameters_from_file.)

Calculating the Hessian

By default, estimate will recompute the Hessian matrix. You can skip calculation of the Hessian matrix entirely if you provide a file with a Hessian that has been pre-computed.

m = Model990()
specify_mode!(m, "path/to/parameter/mode/file.h5")
specify_hessian(m, "path/to/Hessian/matrix/file.h5")
estimate(m)

The specify_hessian function will cause estimate to read in the Hessian matrix rather than calculating it directly. Ensure that you supply an HDF5 file with a variable named hessian that is the correct dimension and data type. Specifying the Hessian matrix but not the parameter mode results in undefined behavior.

See [Hessian Approximation] for more details on the Hessian computation.

Estimation routines

Prior, Likelihood and Posterior calculations

DSGE.prior — Function.

prior(m::AbstractModel{T})

Calculates log joint prior density of m.parameters.

DSGE.likelihood — Function.

likelihood{T<:AbstractFloat}(m::AbstractModel, data::Matrix{T};
                              mh::Bool = false, catch_errors::Bool = false)

Evaluate the DSGE likelihood function. Can handle "two part" estimation where the observed sample contains both a normal stretch of time (in which interest rates are positive) and a stretch of time in which interest rates reach the zero lower bound. If there is a zero-lower-bound period, then we filter over the 2 periods separately. Otherwise, we filter over the main sample all at once.

Arguments

m: The model object

data: matrix of data for observables

Optional Arguments

mh: Whether metropolis_hastings is the caller. If mh=true, the transition matrices for the zero-lower-bound period are returned in a dictionary.

catch_errors: If mh = true, GensysErrors should always be caught.

DSGE.posterior — Function.

posterior{T<:AbstractFloat}(m::AbstractModel{T}, data::Matrix{T};
                             mh::Bool = false, catch_errors::Bool = false)

Calculates and returns the log of the posterior distribution for m.parameters:

log posterior = log likelihood + log prior
log Pr(Θ|data)  = log Pr(data|Θ)   + log Pr(Θ)

Arguments

m: the model object

data: matrix of data for observables

Optional Arguments

-mh: Whether metropolis_hastings is the caller. If mh=true, the log likelihood and the transition matrices for the zero-lower-bound period are also returned. -catch_errors: Whether or not to catch errors of type GensysError or ParamBoundsError

DSGE.posterior! — Function.

posterior!{T<:AbstractFloat}(m::AbstractModel{T}, parameters::Vector{T}, data::Matrix{T};
                              mh::Bool = false, catch_errors::Bool = false)

Evaluates the log posterior density at parameters.

Arguments

m: The model object

parameters: New values for the model parameters

data: Matrix of input data for observables

Optional Arguments

mh: Whether metropolis_hastings is the caller. If mh=true, the log likelihood and the transition matrices for the zero-lower-bound period are also returned.

catch_errors: Whether or not to catch errors of type GensysError or ParamBoundsError. If mh = true, both should always be caught.

Optimization

See Optimization

Full Estimation Routine

See estimate

Output Analysis

DSGE.compute_moments — Function.

compute_moments(m::AbstractModel, percent::Float64 = 0.90; verbose::Symbol=:none)

Computes prior and posterior parameter moments. Tabulates prior mean, posterior mean, and bands in various LaTeX tables stored tablespath(m).

Arguments

m: the model object

percent: the percentage of the mass of draws from Metropolis-Hastings included between the bands displayed in output tables.

DSGE.find_density_bands — Method.

find_density_bands(draws::Matrix, percent::AbstractFloat; minimize::Bool=true)

Returns a 2 x cols(draws) matrix bands such that percent of the mass of draws[:,i] is above bands[1,i] and below bands[2,i].

Arguments

draws: Matrix of parameter draws (from Metropolis-Hastings, for example)

percent: percent of data within bands (e.g. .9 to get 90% of mass within bands)

Optional Arguments

minimize: if true, choose shortest interval, otherwise just chop off lowest and highest (percent/2)

DSGE.make_moment_tables — Method.

make_moment_tables{T<:AbstractFloat}(m::AbstractModel, draws::Matrix{T},
    percent::AbstractFloat; verbose::Symbol = :none)

Tabulates parameter moments in 3 LaTeX tables:

For MAIN parameters, a list of prior means, prior standard deviations, posterior means, and 90% bands for posterior draws

For LESS IMPORTANT parameters, a list of the prior means, prior standard deviations, posterior means and 90% bands for posterior draws.

A list of prior means and posterior means

Arguments

draws: [n_draws x n_parameters] matrix holding the posterior draws from Metropolis-Hastings from save/mhsave.h5

percent: the mass of observations we want; 0 <= percent <= 1

DSGE.save_mean_parameters — Method.

save_mean_parameters{T<:AbstractFloat}(m::AbstractModel, draws::Matrix{T})

Computes and saves the posterior mean of the parameters.

Arguments

m

draws: n_draws x n_parameters matrix holding the posterior draws from Metropolis-Hastings