Impulse Response Functions

Impulse Responses

The forecast step allows the user to automatically compute impulse responses, but for some purposes, a user may just want impulse responses without having to compute any other quantities. We provide this functionality with the impulse_responses function:

DSGE.impulse_responses — Function.

impulse_responses(m, system)

impulse_responses(system, horizon)

Compute impulse responses for a single draw.

Inputs

m::AbstractDSGEModel: model object

system::System{S}: state-space system matrices

horizon::Int: number of periods ahead to forecast

flip_shocks::Bool: Whether to compute IRFs in response to a positive shock (by default the shock magnitude is a negative 1 std. shock)

where S<:AbstractFloat

Outputs

states::Array{S, 3}: matrix of size nstates x horizon x nshocks of state impulse response functions

obs::Array{S, 3}: matrix of size nobs x horizon x nshocks of observable impulse response functions

pseudo::Array{S, 3}: matrix of size npseudo x horizon x nshocks of pseudo-observable impulse response functions

source

We overload impulse_responses to cover specific use cases. For any AbstractDSGEModel, we can compute the impulse responses over a specified horizon for all endogenous state variables, observables, and pseudo-observables by running

m = AnSchorfheide()
system = compute_system(m)
horizon = 40
states_irf, obs_irf, pseudo_irf = impulse_response(system, horizon)
``

For an `AbstractRepModel` (a subtype of `AbstractDSGEModel` for representative
agent models), we can also grab the impulse responses by running

julia statesirf, obsirf, pseudoirf = impulseresponse(m, system)

This use case requires the user to add a setting
under the key `:impulse_response_horizons`, which is
set by default to 40.

If a user wants to specify a subset of the exogenous shocks
and the size of those shocks, the user can run

julia shocknames = [:gsh, :bsh] shockvalues = [1.0, 1.0] impulseresponses(m, system, horizon, shocknames, shock_values)

For the response of an endogenous state or observable to a specific shock,

julia shockname = :gsh varname = :obsgdp varvalue = 0. impulseresponses(m, system, horizon, shockname , varname, var_value) ```