Computing Means and Bands

Procedure

After running a full-distribution forecast, we are often interested in finding means and density bands of the various forecast outputs. This will allow us to plot our estimation of the full distribution of the forecast outputs.

Main Steps:

Load data: Load and transform data, population data, and population forecast (required for forecast output transformations)

Read in forecast outputs: Read in the outputs saved by forecast_one, one variable (e.g. one observable) and one output_type at a time.

Transform forecast outputs: If necessary, use reverse_transform to apply transformations specified in the Observable or PseudoObservable type to the given forecast output series.

Compute means and bands: Compute the means and density bands of the forecast output.

Write to file: For each output_var, Write a MeansBands object (see The MeansBands Type below) to the file specified by get_meansbands_output_files(m).

Computing means and bands is done by calling means_bands_all. If desired, you can also write your computed means and bands as matrices by calling meansbands_matrix_all.

For example, to compute means and bands for an unconditional, full-distribution forecast of states and observables:

m = AnSchorfheide()
means_bands_all(m, :mode, :none, [:forecaststates, forecastobs])

DSGE.means_bands_all — Function.

means_bands_all(m, input_type, cond_type, output_vars; forecast_string = "",
    density_bands = [0.5, 0.6, 0.7, 0.8, 0.9], minimize = false, verbose = :low)

means_bands_all(input_type, cond_type, output_vars, input_dir, output_dir,
    filestring_base; forecast_string = "", density_bands = [0.5, 0.6, 0.7, 0.8, 0.9],
    minimize = false, population_mnemonic = Nullable{Symbol}(),
    population_data_file = "", population_forecast_file = "",
    y0_indexes = Dict{Symbol, Int}(), data = Matrix{T}(),
    verbose = :low)

Computes means and bands for pseudo-observables, observables, and shocks, and writes the results to a file. Two methods are provided. The method that accepts a model object as an argument uses the model's settings to infer the arguments to the second method. Users can optionally skip construction of a model object and manually enter the directory and file names.

Below, T<:AbstractFloat:

Input Arguments

m::AbstractModel: model object

input_type::Symbol: see forecast_one

cond_type::Symbol: see forecast_one

output_vars::Vector{Symbol}: see forecast_one

Method 2 only:

input_dir::AbstractString: directory from which the forecast outputs are to be read in

output_dir::AbstractString: directory to which the means and bands are to be saved

filestring_base::Vector{ASCIIString}: should be equivalent to the result of filestring_base(m)

Keyword Arguments

forecast_string::AbstractString: forecast identifier string (the value "fcid=value" in the forecast output filename). Required when input_type == :subset

density_bands::Vector{T}: a vector of percent values (between 0 and 1) for which to compute density bands

minimize::Bool: if true, choose shortest interval, otherwise just chop off lowest and highest (percent/2)

verbose: level of error messages to be printed to screen. One of :none, :low, :high

Method 2 only:

population_mnemonic::Nullable{Symbol}: the result of get_setting(m, :population_mnemonic. Typically Nullable(:CNP16OV__FRED).

population_data_file::AbstractString: path to population data (in levels) file. In the first method, the following file is used, if it exists: inpath(m, "data", "population_data_levels_(data_vintage(m)).csv")

population_forecast_file::S: path to population forecast (in levels) file. In the first method, if use_population_forecast(m), the following file is used, if it exists: inpath(m, "data", "population_forecast_(data_vintage(m)).csv")

y0_indexes::Dict{Symbol, Int}: A Dict storing the mapping of products to the index of the first period of data needed to compute the product's transformation. This is used to compute growth rates and four-quarter cumulations (among others). See get_y0_index.

data::Matrix{T}: pre-loaded nobs x nperiods matrix containing the (untransformed) data matrix.

The MeansBands Type

DSGE.MeansBands — Type.

type MeansBands

Stores the means and bands of results for a particular set of outputs from the forecast step.

Specifically, forecasts can be made for any element in the Cartesian product of 4 sets:

input_type: some subset of the parameter draws from the estimation step. See forecast_one for all possible options.

cond_type: conditional type. See forecast_one for all possible options.

product: a particular result computed in the forecast. This could be one of the following:

  - `hist`: smoothed histories
  - `forecast`: forecasted values
  - `shockdec`: shock decompositions
  - `irf`: impulse responses

variable class: the category in which a particular variable, like :y_t, falls. Options are:

  - `state`: state (from `m.endogenous_states` or `m.endogenous_states_augmented`)
  - `obs`: observable (from `m.observables`)
  - `pseudo`: pseudoobservable (from `pseudo_measurement` equation)
  - `shock`: shock (from `m.exogenous_shocks`)

Note that the Cartesian product (product x class) is the set of options for output_vars in the forecast_one function signature.

Fields

metadata::Dict{Symbol,Any}: Contains metadata keeping track of the input_type, cond_type, product (history, forecast, shockdec, etc), and variable class (observable, pseudoobservable, state, etc) stored in this MeansBands structure.

means::DataFrame: a DataFrame of the mean of the time series

bands::Dict{Symbol,DataFrame}: a Dict mapping variable names to DataFrames containing confidence bands for each variable. See find_density_bands for more information.