Special Model Types

Special Model Types

In addition to the AbstractDSGEModel type, DSGE.jl* has several special types that implement models which are designed to interface with DSGEs.

The PoolModel Type

Unlike the other models contained in DSGE, the PoolModel type is not a proper DSGE model. It is a wrapper object for different methods to perform model averaging for two different models, which do not have to be DSGE models. For example, a user could average two different vector auto-regressions. Generally, a user only needs to provide the predictive density scores of the two models that the user wants to average. The reason is that we treat the predictive density scores as non-FRED observables. This approach makes interfacing with the rest of the machinery provided by DSGE.jl* very simple.

DSGE.PoolModel — Type.

PoolModel{T}

Implements several model averaging methods to pool the predictions of different models. Currently, PoolModel only works for pooling two different models, although it can be extended to more models. Confer with the paper for details.

The available methods are

dynamic: weights on pooled models evolve over time and are treated as a stochastic process.

static: weights on pooled models are assumed time-invariant.

equal: weights are set to 1/2.

bma: weights are updated according to Bayesian model averaging.

The default method is dynamic, which implements the Dynamic Pools method developed by Del Negro et al. (2016). To choose another method, use the keyword weight_type::Symbol, e.g.

pm = PoolModel(weight_type = :static) # creates a static PoolModel

The weight_type is stored as a setting, so users can retrieve at any point by using get_setting.

Fields

Parameters

parameters::Vector{AbstractParameter}: Vector of all time-invariant hyperparameters for the chosen method of model averaging.

keys::OrderedDict{Symbol,Int}: Maps human-readable names for predictive densities of pooled models.

Inputs to Measurement and Equilibrium Condition Equations

model::OrderedDict{Symbol,AbstractDSGEModel}: Maps name to its underlying model object.

Model Specifications and Settings

spec::String: The model specification identifier, "an_schorfheide", cached here for filepath computation.

subspec::String: The model subspecification number, indicating that some parameters from the original model spec ("ss0") are initialized differently. Cached here for filepath computation.

settings::Dict{Symbol,Setting}: Settings/flags that affect computation without changing the economic or mathematical setup of the model.

test_settings::Dict{Symbol,Setting}: Settings/flags for testing mode

Other Fields

rng::MersenneTwister: Random number generator. Can be is seeded to ensure reproducibility in algorithms that involve randomness (such as Metropolis-Hastings).

testing::Bool: Indicates whether the model is in testing mode. If true, settings from m.test_settings are used in place of those in m.settings.

observable_mappings::OrderedDict{Symbol,Observable}: A dictionary that stores data sources, series mnemonics, and transformations to/from model units. DSGE.jl will fetch data from the Federal Reserve Bank of St. Louis's FRED database; all other data must be downloaded by the user. See load_data and Observable for further details.

source

See Del Negro et al. (2016) for theoretical details on the model averaging methods listed in the documentation.

To facilitate analysis with the PoolModel type, we also provide the following functions.

DSGE.estimate_bma — Function.

estimate_bma(m, data, prior; return_output = false, filestring_addl = [])

Estimate a Bayesian Model Average.

Arguments:

m::PoolModel: PoolModel object

Optional Arguments:

data: well-formed data as Matrix or DataFrame. If this is not provided, the load_data routine will be executed.

prior::Float64: prior probability between the two models

Keyword Arguments:

return_output::Bool: option to return output. If false, nothing is returned.

filestring_addl::Vector{String}: Additional strings to append to output files.

source

DSGE.sample_λ — Function.

sample_λ(m, pred_dens, θs, T = -1; parallel = true) where S<:AbstractFloat
sample_λ(m, pred_dens, T = -1; parallel = true) where S<:AbstractFloat

Computes and samples from the conditional density p(λt|θ, It, P) for particle in θs, which represents the posterior distribution. The sampled λ particles represent the posterior distribution p(λ{t|t} | It, P).

If no posterior distribution is passed in, then the function computes the distribution of λ_{t|t} for a static pool.

Inputs

m::PoolModel{S}: PoolModel object

pred_dens::Matrix{S}: matrix of predictive densities

θs::Matrix{S}: matrix of particles representing posterior distribution of θ

T::Int64: final period for tempered particle filter

where S<:AbstractFloat.

Keyword Argument

parallel::Bool: use parallel computing to compute and sample draws of λ

Outputs

λ_sample::Vector{Float64}: sample of draws of λs; together with (θ,λ) represents a joint density

```

source

DSGE.propagate_λ — Function.

propgate_λ(λvec, h, m, θvec) where T<:AbstractFloat

Propagates a λ particle h periods forward.

Inputs

λ::T: λ sample from (θ,λ) joint distribution

h::Int64: forecast horizon

m::PoolModel: PoolModel object

θvec::Vector{T}: optional vector of parameters to update PoolModel

```

source

DSGE.compute_Eλ — Function.

compute_Eλ(m, h, λvec, θmat = [], weights = [];
    current_period = true, parallel = true) where T<:AbstractFloat

Computes and samples from the conditional density p(λt|θ, It, P) for particle in θs, which represents the posterior distribution.

Inputs

m::PoolModel{T}: PoolModel object

h::Int64: forecast horizon

λvec::Vector{T}: vector of particles of λ samples from (θ,λ) joint distribution

`θmat::Matrix{T}': matrix of posterior parameter samples

weights::Vector{T}: weights of λ particles, defaults to equal weights

Keyword Argument

current_period::Bool: compute Eλ for current period t

parallel::Bool: use parallel computing to compute and sample λ

get_dpp_pred_dens::Bool: compute predictive densities according to dynamic prediction pools

Outputs

λhat_tplush::Float64: E[λ{t+h|t} | It^P, P]

λhat_t::Float64: E[λ{t|t} | It^P, P]

```

source

DSGE-VARs and the DSGEVAR Type

We can approximate the dynamics of a linearized DSGE with a VAR($p$), where $p$ is the number of lags. This approximation gives a mapping from the parameters of a DSGE to the parameters of a VAR (the coefficients and innovations variance-covariance matrix). Since the number of parameters in a VAR are generally larger than the number of parameters in a DSGE, this mapping can be interpreted as cross-restrictions imposed by a DSGE on the parameters of a VAR. A DSGE-VAR combines a DSGE with a VAR to, among other reasons, evaluate the mis-specification of the DSGE and improve the DSGE's forecasting performance.

For more details on the theory, see Del Negro and Schorfheide (2004), Del Negro and Schorfheide (2006), and Del Negro and Schorfheide (2009).

We implement DSGE-VARs with the DSGEVAR concrete type so that it is easy to interface them with DSGE models. A DSGEVAR type holds information about the VAR with which we want to combine a given DSGE model and can be easily constructed, given a DSGE object. Once we have constructed a DSGEVAR object, then it is straightoforward to estimate the object and compute impulse responses.

DSGE.DSGEVAR — Type.

DSGEVAR{T} <: AbstractDSGEVARModel{T}

implements a simple interface for combining a given DSGE model with a VAR to create a DSGE-VAR. Confer with Del Negro and Schorfheide (2004), Del Negro and Schorfheide (2006), and Del Negro and Schorfheide (2009) for details about DSGE-VARs.

The recommended constructor requires the user to provide (1) an AbstractDSGEModel object, (2) which structural shocks from the DSGE to use, and (3) the subspec (optional, defaults to "ss0"). If the subspec "ss0" is used, then the result is a DSGEVAR whose VAR component is "empty" in that the observables, lags, and λ weight are not specified. The reason why this constructor requires the user to specify which structural shocks of DSGE to use is that this information is DSGE-specific rather than information about the VAR.

However, we can also construct a DSGEVAR without having to specify the structural shocks when calling the constructor, although we still need to give an instance of an AbstractDSGEModel.

Example

The code below instantiates an empty DSGEVAR with AnSchorfheide as the underlying DSGE and then calls update! on the empty DSGE-VAR to add information about the desired DSGE-VAR spec.

dsgevar = DSGEVAR(AnSchorfheide())
DSGE.update!(dsgevar, shocks = [:rm_sh, :z_sh, :g_sh],
    observables = [:obs_gdp, :obs_cpi, :obs_nominalrate],
    λ = 1., lags = 4)

Fields

DSGE object

dsge::AbstractDSGEModel{T}: underlying DSGE model object

DSGE-VAR Information

observables::OrderedDict{Symbol,Int}: dictionary mapping observables of the VAR to their index in the matrices representing the DSGE-VAR

shocks::OrderedDict{Symbol,Int}: dictionary mapping structural shocks in the DSGE to their index in the matrices representing the DSGE-VAR

lags::Int: number of lags in the VAR

λ::T: weight on the DSGE prior

Auxiliary Information

spec::String: concatenates dsgevar with the spec of the DSGE, e.g. for AnSchorfheide, we have dsgevar_an_schorfheide.

subspec::String: specifies the model subspecification. Cached here for filepath computation.

testing::Bool: indicates whether the model is in testing mode. Currently, this setting has no uses in practice

source

Tips for Using DSGEVARs

When extensively using DSGE-VARs, we recommend defining your own subspecs in subspecs.jl because it simplifies the process of saving, reading, and analyzing output from estimating and calculating impulse responses for DSGE-VARs. See Advanced Usage for a more detailed explanation on changing subspecs.

The names of the observables must exist as either observables or pseudo-observables in the DSGE because for most DSGEVAR methods, we need to construct the state space representation of the DSGEVAR using information from the underlying DSGE.

It is important to be careful about the order of the observables when constructing a DSGEVAR. Whether you define the names of the observables by calling update! or by creating a subspec, we assume that the order of the observables corresponds to the observable's index in the data and in the state space representation of the DSGEVAR. In the example provided above, if we estimate the DSGEVAR on data or construct the state space representation of the DSGEVAR, we assume that the order of observables in the data array, which has dimensions nobs x nperiods, is :obs_gdp in the first row, :obs_cpi in the second row, and :obs_nominalrate in the third row.

When using functions that use the DSGE as a prior for a VAR (as opposed to a VAR approximation of the DSGE), then an intercept term is assumed and cannot be turned off. For example, the following two functions computes the VAR coefficients and innovations variance-covariance matrix for a DSGEVAR object m. The first one is for a VAR approximation of the DSGE in m, and it allows the user to specify whether or not they want an intercept term using the keyword use_intercept. The second function is for using the DSGE as a prior for a VAR estimated on data. This function does not have the use_intercept keyword because we require an intercept term when using the DSGE as a prior for a VAR.

compute_system(m; apply_altpolicy = false,
               regime_switching = false, n_regimes = 2,
               check_system = false, get_system = false,
               get_population_moments = false, use_intercept = false,
               verbose = :high)
compute_system(m, data; apply_altpolicy = false,
               regime_switching = false, n_regimes = 2,
               check_system = false, get_system = false,
               get_population_moments = false,
               verbose = :high)