Alternative Policies

Procedure for Permanent Alternative Policies

This section describes forecasting under a (permanent) alternative monetary policy rule. That is:

Filtering and smoothing is done under the historical monetary policy rule, i.e. the one defined in the eqcond method for the given model.

Before forecasting, the state space matrices are recomputed under the alternative policy rule.

Forecasts and IRFs are computed under the alternative rule.

Note that shock decompositions (and the two associated products, trends and deterministic trends) cannot currently be computed under an alternative policy.

The user defines some instance of the AltPolicy type (described below) and sets it as the value for the :alternative_policy Setting. Then the function calls made to forecast and compute means and bands remain the same as usual (see Forecasting and Computing Means and Bands).

For example, suppose you have defined the functions taylor93_eqcond and taylor93_solve corresponding to Taylor (1993)'s proposed monetary policy rule. Then you can run:

m = AnSchorfheide()
m <= Setting(:alternative_policy, AltPolicy(:taylor93, taylor93_eqcond, taylor93_solve))
forecast_one(m, :mode, :none, [:forecastobs, :forecastpseudo])
compute_meansbands(m, :mode, :none, [:forecastobs, :forecastpseudo])

Procedure for Temporary Alternative Policies

Another counterfactual exercise is temporarily imposing a different monetary policy rule, i.e. a temporary alternative policy. To implement this, we utilize exogenous regime switching in the forecast horizon. See Regime-Switching Forecasts for details on regime-switching.

In a rational expectations equilibrium, agents will take into account the fact that the temporary policy is expected to terminate. A different algorithm than Chris Sims's standard gensys algorithm is required, which we have implemented as gensys_cplus.

To set up a temporary alternative policy, a user needs to specify the changes to the equilibrium conditions to the policy rule in eqcond For instance, a Nominal GDP targeting policy. uses the function ngdp_replace_eq_defines entries to define these changes.

Missing docstring.

Missing docstring for DSGE.ngdp_replace_eq_entries. Check Documenter's build log for details.

The user also needs to complete the following steps.

Adding a regime for every period in the forecast horizon during which the alternative policy applies, plus one more regime for the first regime in which the alternative policy does NOT apply.

Adding the setting Setting(:gensys2, true) to indicate gensys_cplus should be used

Adding the setting Setting(:replace_eqcond, true) to indicate the eqcond function will be replaced

Adding the setting Setting(:replace_eqcond_func_dict, replace_eqcond), where replace_eqcond should be a Dict{Int, Function} mapping regimes to alternative eqcond functions. Note that the user only needs to populate regimes in which the eqcond function differs from the standard one.

To see an example of using temporary alternative policies, see the example script for regime-switching.

In contrast, the permanent version of Nominal GDP targeting would be

AltPolicy(policy, DSGE.ngdp_eqcond, DSGE.ngdp_solve, forecast_init = DSGE.ngdp_forecast_init)

where

DSGE.ngdp_eqcond — Function

ngdp_eqcond(m::AbstractDSGEModel)

Solves for the transition equation of m under a price level targeting rule (implemented by adding a price-gap state)

DSGE.ngdp_solve — Function

ngdp_solve(m::AbstractDSGEModel)

Solves for the transition equation of m under a price level targeting rule (implemented by adding a price-gap state)

DSGE.ngdp_forecast_init — Function

init_ngdp_forecast(m::AbstractDSGEModel, shocks::Matrix{T}, final_state::Vector{T})

Adjust shocks matrix and final state vector for forecasting under the NGDP rule

The AltPolicy Type

DSGE.AltPolicy — Type

mutable struct AltPolicy

Type defining an alternative policy rule.

Fields

key::Symbol: alternative policy identifier

eqcond::Function: a version of DSGE.eqcond which computes the equilibrium condition matrices under the alternative policy. Like DSGE.eqcond, it should take in one argument of mutable struct AbstractDSGEModel and return the Γ0, Γ1, C, Ψ, and Π matrices.

solve::Function: a version of DSGE.solve which solves the model under the alternative policy. Like DSGE.solve, it should take in one argument of mutable struct AbstractDSGEModel and return the TTT, RRR, and CCC matrices.

forecast_init::Function: a function that initializes forecasts under the alternative policy rule. Specifically, it accepts a model, an nshocks x n_forecast_periods matrix of shocks to be applied in the forecast, and a vector of initial states for the forecast. It must return a new matrix of shocks and a new initial state vector. If no adjustments to shocks or initial state vectors are necessary under the policy rule, this field may be omitted.

color::Colorant: color to plot this alternative policy in. Defaults to blue.

linestyle::Symbol: line style for forecast plots under this alternative policy. See options from Plots.jl. Defaults to :solid.