Running An Existing Model

Running an Existing Model

The DSGE.jl package provides several example models:

A simple three-equation DSGE model from An and Schorfheide (2006)

The well-known Smets and Wouters (2007) model

The New York Fed DSGE model (version 990.2), which was introduced in this blog post

The New York Fed DSGE model (version 1002.9), which is documented here

The New York Fed DSGE model (version 1010.18)

Several methods for model averaging over two models from Del Negro et al. (2016) through the concrete type PoolModel.

You can run these models using the description provided here. If you were to implement another model using DSGE.jl, these procedures can also be used to estimate those models.

Running with Default Settings

To estimate and forecast in Julia, simply create an instance of the model object and call estimate and forecast_all. A minimal example is reproduced below:

# estimate as of 2015-Q3 using the default data vintage from 2015 Nov 27
custom_settings = Dict{Symbol, Setting}(
    :data_vintage        => Setting(:data_vintage, "151127"),
    :date_forecast_start => Setting(:date_forecast_start, quartertodate("2015-Q4")))

# construct a model object
m = Model990(custom_settings = custom_settings)

# reoptimize parameter vector, compute Hessian at mode, and full posterior
# parameter sampling
estimate(m)

# produce LaTeX tables of parameter moments
compute_moments(m)

# forecast and compute means and bands using 10 processes
my_procs = addprocs(10)
@everywhere using DSGE

forecast_one(m, :full, :none, [:forecaststates, :forecastobs])
compute_meansbands(m, :full, :none, [:forecaststates, :forecastobs])
rmprocs(my_procs)

For more details on changing the model's default settings, parameters, equilibrium conditions, etc., see Advanced Usage.

By default, the estimate routine loads the dataset, reoptimizes the initial parameter vector, computes the Hessian at the mode, and conducts full posterior parameter sampling using Metropolis-Hastings. (The initial parameter vector used is specified in the model's constructor.) Further options for estimation are described in Estimation:

To use updated data or alternative user-specified datasets, see Input Data.

The user may want to avoid reoptimizing the parameter vector and calculating the Hessian matrix at this new vector. Please see Reoptimizing.

For more information on the many types of forecasts that can be run on an existing or user-defined model, see Forecasting.

Input/Output Directory Structure

The DSGE.jl estimation uses data files as input and produces large data files as outputs. One estimation saves several GB of parameter draws and related outputs. It is useful to understand how these files are loaded/saved and how to control this behavior.

Directory Tree

The following subdirectory tree indicates the default locations of these input and outputs. Square brackets indicate directories in the tree that will become relevant as future features are implemented.

Note that this directory tree is not linked, although it appears to be.

<dataroot>/: Root data directory.

data/: Macroeconomic input data series.

cond/: Conditional data, i.e.

user/: User-created or sample model input files.

<saveroot>/: Root save directory.

output_data/

m990/: Input/output files for the Model990 type. A model of type SPEC will create its own save directory SPEC/ at this level in the directory tree.

ss0/: Subdirectory for subspec 0. A model of a different subspec will have similar directories at this level of the tree.

estimate/

figures/: Plots and other figures

tables/: LaTeX tables

raw/: Raw output data from estimation step

work/: Derived data files created using raw/ files as input

[xxx/]: Other model outputs, such as forecasts, impulse response functions, and shock decompositions; subdirectory structure mirrors that of estimate.

Directory Paths

By default, input/output directories are located in the DSGE.jl package, along with the source code. Default values of the input/output directory roots:

saveroot(m): "$(Pkg.dir())/DSGE/save"

dataroot(m): "$(Pkg.dir())/DSGE/save/input_data"

Note these locations can be overridden as desired. See Model Settings for more details.

m <= Setting(:saveroot, "path/to/my/save/root")
m <= Setting(:dataroot, "path/to/my/data/root")

Utility functions are provided to create paths to input/output files. These should be used for best results.

Missing docstring.

Missing docstring for DSGE.inpath. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DSGE.rawpath. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DSGE.logpath. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DSGE.workpath. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DSGE.tablespath. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DSGE.figurespath. Check Documenter's build log for details.

The PoolModel Type

Unlike the other models contained in DSGE, the PoolModel type is not a proper DSGE model. It is a wrapper object for different methods to average two different models, which do not have to be DSGE models. For example, a user could average two different vector auto-regressions. Generally, a user only needs to provide the predictive density scores of the two models that the user wants to average. The reason is that we treat the predictive density scores as non-FRED observables. This approach makes interfacing with the rest of the machinery provided by DSGE.jl very simple.

DSGE.PoolModel — Type.

PoolModel{T}

Implements several model averaging methods to pool the predictions of different models. Currently, PoolModel only works for pooling two different models, although it can be extended to more models. Confer with the paper for details.

The available methods are

dynamic: weights on pooled models evolve over time and are treated as a stochastic process.

static: weights on pooled models are assumed time-invariant.

equal: weights are set to 1/2.

bma: weights are updated according to Bayesian model averaging.

The default method is dynamic, which implements the Dynamic Pools method developed by Del Negro et al. (2016). To choose another method, use the keyword weight_type::Symbol, e.g.

pm = PoolModel(weight_type = :static) # creates a static PoolModel

The weight_type is stored as a setting, so users can retrieve at any point by using get_setting.

Fields

Parameters

parameters::Vector{AbstractParameter}: Vector of all time-invariant hyperparameters for the chosen method of model averaging.

keys::OrderedDict{Symbol,Int}: Maps human-readable names for predictive densities of pooled models.

Inputs to Measurement and Equilibrium Condition Equations

model::OrderedDict{Symbol,AbstractDSGEModel}: Maps name to its underlying model object.

Model Specifications and Settings

spec::String: The model specification identifier, "an_schorfheide", cached here for filepath computation.

subspec::String: The model subspecification number, indicating that some parameters from the original model spec ("ss0") are initialized differently. Cached here for filepath computation.

settings::Dict{Symbol,Setting}: Settings/flags that affect computation without changing the economic or mathematical setup of the model.

test_settings::Dict{Symbol,Setting}: Settings/flags for testing mode

Other Fields

rng::MersenneTwister: Random number generator. Can be is seeded to ensure reproducibility in algorithms that involve randomness (such as Metropolis-Hastings).

testing::Bool: Indicates whether the model is in testing mode. If true, settings from m.test_settings are used in place of those in m.settings.

observable_mappings::OrderedDict{Symbol,Observable}: A dictionary that stores data sources, series mnemonics, and transformations to/from model units. DSGE.jl will fetch data from the Federal Reserve Bank of St. Louis's FRED database; all other data must be downloaded by the user. See load_data and Observable for further details.

source

See Del Negro et al. (2016) for theoretical details on the model averaging methods listed in the documentation.

To facilitate analysis with the PoolModel type, we also provide the following functions.

DSGE.estimate_bma — Function.

estimate_bma(m, data, prior; return_output = false, filestring_addl = [])

Estimate a Bayesian Model Average.

Arguments:

m::PoolModel: PoolModel object

Optional Arguments:

data: well-formed data as Matrix or DataFrame. If this is not provided, the load_data routine will be executed.

prior::Float64: prior probability between the two models

Keyword Arguments:

return_output::Bool: option to return output. If false, nothing is returned.

filestring_addl::Vector{String}: Additional strings to append to output files.

source

DSGE.sample_λ — Function.

sample_λ(m, pred_dens, θs, T = -1; parallel = true) where S<:AbstractFloat
sample_λ(m, pred_dens, T = -1; parallel = true) where S<:AbstractFloat

Computes and samples from the conditional density p(λt|θ, It, P) for particle in θs, which represents the posterior distribution. The sampled λ particles represent the posterior distribution p(λ{t|t} | It, P).

If no posterior distribution is passed in, then the function computes the distribution of λ_{t|t} for a static pool.

Inputs

m::PoolModel{S}: PoolModel object

pred_dens::Matrix{S}: matrix of predictive densities

θs::Matrix{S}: matrix of particles representing posterior distribution of θ

T::Int64: final period for tempered particle filter

where S<:AbstractFloat.

Keyword Argument

parallel::Bool: use parallel computing to compute and sample draws of λ

Outputs

λ_sample::Vector{Float64}: sample of draws of λs; together with (θ,λ) represents a joint density

```

source

Missing docstring.

Missing docstring for propagate_λ. Check Documenter's build log for details.

DSGE.compute_Eλ — Function.

compute_Eλ(m, h, λvec, θmat = [], weights = [];
    current_period = true, parallel = true) where T<:AbstractFloat

Computes and samples from the conditional density p(λt|θ, It, P) for particle in θs, which represents the posterior distribution.

Inputs

m::PoolModel{T}: PoolModel object

h::Int64: forecast horizon

λvec::Vector{T}: vector of particles of λ samples from (θ,λ) joint distribution

`θmat::Matrix{T}': matrix of posterior parameter samples

weights::Vector{T}: weights of λ particles, defaults to equal weights

Keyword Argument

current_period::Bool: compute Eλ for current period t

parallel::Bool: use parallel computing to compute and sample λ

get_dpp_pred_dens::Bool: compute predictive densities according to dynamic prediction pools

Outputs

λhat_tplush::Float64: E[λ{t+h|t} | It^P, P]

λhat_t::Float64: E[λ{t|t} | It^P, P]

```

source